Se implementó el modo offline de cleintes sin direcciones

+++ lib/services/CustomerService.dart
@@ -30,7 +30,7 @@ Future<ResponseModel> getAllCustomers(String customer, String authorization, { S
 
   CustomersModel customersObj = new CustomersModel(data: customers, perPage: 0);
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: customersObj.toJson());
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customersObj);
 
   return response;
 }
@@ -56,7 +56,7 @@ Future<ResponseModel> getAllCustomersWithAddress(String customer, String authori
 
   CustomersWithAddressModel customersObj = new CustomersWithAddressModel(data: customers, perPage: 0);
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: customersObj.toJson());
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customersObj);
 
   return response;
 }
@@ -82,7 +82,7 @@ Future<ResponseModel> getCustomer(String id, String customer, String authorizati
 
   CustomerModel customerObj = await DatabaseProvider.db.ReadCustomerById(int.parse(id));
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: customerObj.toJson());
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customerObj);
 
   return response;
 }
@@ -99,15 +99,16 @@ Future<ResponseModel> createCustomer(CustomerModel customerObj, String customer,
   if (isOnline) {
     var createCustomerResponse = await createCustomerFromServer(customerObj, customer, authorization);
     if (createCustomerResponse.statusCode==200 || createCustomerResponse.statusCode==201) {
-      var a = createCustomerResponse.body;
       customerObj = CustomerModel.fromJson(createCustomerResponse.body);
       syncState = SyncState.synchronized;
     }
   }
   
   CustomerModel customerCreated = await DatabaseProvider.db.CreateCustomer(customerObj, syncState);
+  UserModel user = await DatabaseProvider.db.RetrieveLastLoggedUser();
+  await DatabaseProvider.db.CreateCustomerUser(null, null, null, null, customerCreated.id, user.id, syncState);
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: customerCreated.toJson());
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customerCreated);
 
   return response;
 }
@@ -119,11 +120,21 @@ Future<http.Response> createCustomerFromServer(CustomerModel customerObj, String
   return await httpPost(bodyJson, customer, authorization, resourcePath);
 }
 
-Future<ResponseModel> updateCustomer(String id, CustomerModel customerObj, String customer, String authorization) async{
-  
-  CustomerModel customerUpdated = await DatabaseProvider.db.UpdateCustomer(int.parse(id), customerObj, SyncState.updated);
+Future<ResponseModel> updateCustomer(String id, CustomerModel customerObj, String customer, String authorization) async {
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: customerUpdated.toJson());
+  var syncState = SyncState.updated;
+
+  if (isOnline) {
+    var updateCustomerResponse = await updateCustomerFromServer(customerObj.id.toString(), customerObj, customer, authorization);
+    if (updateCustomerResponse.statusCode==200 || updateCustomerResponse.statusCode==201) {
+      customerObj = CustomerModel.fromJson(updateCustomerResponse.body);
+      syncState = SyncState.synchronized;
+    }
+  }
+  
+  CustomerModel customerUpdated = await DatabaseProvider.db.UpdateCustomer(int.parse(id), customerObj, syncState);
+
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customerUpdated);
 
   return response;
 }
@@ -137,7 +148,22 @@ Future<http.Response> updateCustomerFromServer(String id, CustomerModel customer
 
 Future<ResponseModel> deleteCustomer(String id, String customer, String authorization) async {
 
-  int responseDelete = await DatabaseProvider.db.DeleteCustomerById(int.parse(id));
+  bool deletedFromServer = false;
+
+  if (isOnline) {
+    var deleteCustomerResponse = await deleteCustomerFromServer(id, customer, authorization);
+    if (deleteCustomerResponse.statusCode==200 || deleteCustomerResponse.statusCode==201) {
+      deletedFromServer = true;
+    }
+  }
+
+  int responseDelete;
+
+  if (deletedFromServer) {
+    responseDelete = await DatabaseProvider.db.DeleteCustomerById(int.parse(id));
+  } else {
+    responseDelete = await DatabaseProvider.db.ChangeSyncStateCustomer(int.parse(id), SyncState.deleted);
+  }
 
   ResponseModel response = new ResponseModel(statusCode: 200, body: responseDelete.toString());
 
@@ -153,7 +179,7 @@ Future<http.Response> deleteCustomerFromServer(String id, String customer, Strin
 Future<ResponseModel> getCustomerAddresses(String id, String customer, String authorization ) async{
   List<AddressModel> addresses = await DatabaseProvider.db.RetrieveAddressModelByCustomerId(int.parse(id));
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: json.encode(addresses));
+  ResponseModel response = new ResponseModel(statusCode: 200, body: addresses);
 
   return response;
 }
@@ -166,8 +192,17 @@ Future<http.Response> getCustomerAddressesFromServer(String id, String customer,
 }
 
 Future<ResponseModel> relateCustomerAddress(String idCustomer, String idAddress, String customer, String authorization) async{
+
+  var syncState = SyncState.created;
+
+  if (isOnline) {
+    var relateCustomerAddressResponse = await relateCustomerAddressFromServer(idCustomer, idAddress, customer, authorization);
+    if (relateCustomerAddressResponse.statusCode==200 || relateCustomerAddressResponse.statusCode==201) {
+      syncState = SyncState.synchronized;
+    }
+  }
   
-  var customerAddressCreated = await DatabaseProvider.db.CreateCustomerAddress(null, null, null, null, int.parse(idCustomer), int.parse(idAddress), true, SyncState.created);
+  var customerAddressCreated = await DatabaseProvider.db.CreateCustomerAddress(null, null, null, null, int.parse(idCustomer), int.parse(idAddress), true, syncState);
 
   ResponseModel response = new ResponseModel(statusCode: 200, body: customerAddressCreated.toString());
 
@@ -187,9 +222,25 @@ Future<http.Response> relateCustomerAddressFromServer(String idCustomer, String 
 }
 
 Future<ResponseModel> unrelateCustomerAddress(String idCustomer, String idAddress, String customer, String authorization) async{
-  var responseDelete = await DatabaseProvider.db.DeleteCustomerAddressById(int.parse(idCustomer), int.parse(idAddress));
+  var syncState = SyncState.deleted;
+  bool unrelateFromServer = false;
 
-  ResponseModel response = new ResponseModel(statusCode: 200, body: responseDelete.toString());
+  if (isOnline) {
+    var unrelateCustomerAddressResponse = await unrelateCustomerAddressFromServer(idCustomer, idAddress, customer, authorization);
+    if (unrelateCustomerAddressResponse.statusCode==200 || unrelateCustomerAddressResponse.statusCode==201) {
+      unrelateFromServer = true;
+    }
+  }
+
+  var customerAddressDelete;
+
+  if (unrelateFromServer) {
+    customerAddressDelete = await DatabaseProvider.db.DeleteCustomerAddressById(int.parse(idCustomer), int.parse(idAddress));
+  } else {
+    customerAddressDelete = await DatabaseProvider.db.ChangeSyncStateCustomerAddress(int.parse(idCustomer), int.parse(idAddress), syncState);
+  }
+
+  ResponseModel response = new ResponseModel(statusCode: 200, body: customerAddressDelete.toString());
 
   return response;
 }
